+++ lib/src/cook/process/command/CommonImpl.hpp
@@ -30,6 +30,7 @@ namespace cook { namespace process { namespace command {
         toolchain::KeyValuesMap & kvm_;
         toolchain::TranslatorMap & trans_;
         const Language language_;
+        const std::filesystem::path recipe_path_;
     };
 
 } } }